lots of server creation stuff
